Composite Authorization and Service Accounts
This proposal originated in the desire to provide a more secure and
convenient way of storing service-specific data (such as Glance
images) in end-user Swift accounts. When discussing this with the
Keystone team, the concept of composite authorization was invented.

Following review at the Swift design summit in Paris, the proposal has
evolved as described here. Feedback from Glance and Cinder developers
has been incorporated into the proposal.

Change-Id: I962b94a253c717b7938953343232fb8837b05b55