Large containers (Sharding) - Pivot Ranges.
Here is an initial version of the next sharding POC, it is based off
discussions at the hackathon in Austin and the Tokyo design summit,
plus some initial work I have already put together to see how it would
work.

This approach is similar, but simpler then in the last distributed
prefix tree POC. It will simply split the table at a pivot point,
creating two new containers that hold all the objects. The root
container holds a reference to all the new containers and their range,
so it can short circuit requests as it knows all data will be in the tree's
leaves.

Change-Id: I21766423a07a5b6195656b06db4e590598bf3d41