formpost should allow subprefix-based signature
formpost currently requires that the signature used to validate a file
upload contains the same object_prefix as the object_prefix specified
in the action url of the form.
We propose that the middleware should also accept signatures calculated
with a subprefix of the object_prefix in the action url. Thus, formpost
would accept all uploads to pseudofolders which contain a common
subprefix.
With this, sharing of data with external people is made much easier via
webbased applications, because only one signature is needed in order to
create forms for every pseudofolder in a container.

Change-Id: I4237f141022382131977ff16760f5645d0391ea5